admin-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
sqlalchemy>=2.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
slowapi>=0.1.9
